break out notebook app component from core
