+++ applications/jupyter-extension/nteract_on_jupyter/package.json
@@ -18,8 +18,10 @@
     "@nteract/commutable": "^4.0.1",
     "@nteract/core": "^5.0.1",
     "@nteract/editor": "^7.0.4",
+    "@nteract/notebook-app-component": "^1.0.0",
     "@nteract/notebook-preview": "^7.1.1",
     "@nteract/octicons": "^0.2.0",
+    "@nteract/presentational-components": "^0.0.1",
     "@nteract/timeago": "^3.4.9",
     "@nteract/transform-dataresource": "^3.0.2",
     "@nteract/transform-geojson": "^3.1.0",
