+++ applications/notebook-on-next/pages/edit.js
@@ -3,7 +3,10 @@
 import * as React from "react";
 import fetch from "isomorphic-fetch";
 import { monocellNotebook, fromJS, toJS } from "@nteract/commutable";
-import { NotebookApp, Styles } from "@nteract/core";
+
+import NotebookApp from "@nteract/notebook-app-component";
+import { Styles } from "@nteract/presentational-components";
+
 import { Provider } from "react-redux";
 import { List as ImmutableList, Map as ImmutableMap } from "immutable";
