+++ packages/core/package.json
@@ -16,26 +16,12 @@
   },
   "dependencies": {
     "@nteract/commutable": "^4.0.1",
-    "@nteract/display-area": "^4.1.1",
-    "@nteract/dropdown-menu": "^0.1.2",
-    "@nteract/editor": "^7.0.4",
-    "@nteract/markdown": "^1.1.0",
-    "@nteract/mathjax": "^1.0.0",
     "@nteract/messaging": "^4.0.0",
-    "@nteract/octicons": "^0.2.0",
-    "@nteract/presentational-components": "^0.0.1",
-    "@nteract/styles": "^1.0.1",
-    "@nteract/transforms": "^4.1.1",
-    "@nteract/types": "^2.0.0",
-    "date-fns": "^1.29.0",
     "escape-carriage": "^1.2.0",
     "file-saver": "^1.3.3",
     "prop-types": "^15.6.1",
     "rc-menu": "6.2.10",
-    "react-dnd": "^2.5.4",
-    "react-dnd-html5-backend": "^2.5.4",
     "react-redux": "^5.0.7",
-    "react-syntax-highlighter": "^7.0.1",
     "redux": "^4.0.0",
     "redux-immutable": "^4.0.0",
     "redux-observable": "^0.18.0",
