+++ packages/core/src/components/index.js
@@ -1,11 +1,10 @@
 // @flow
 
-export * from "@nteract/presentational-components";
 export {
   default as NewNotebookNavigation,
   PureNewNotebookNavigation
 } from "./new-notebook-navigation";
+
 export { default as NotebookMenu } from "./notebook-menu";
 export { default as TitleBar, TitleBar as PureTitleBar } from "./title-bar";
 export { default as ModalController } from "./modal-controller";
-export { default as NotebookApp } from "./notebook-app";
