+++ packages/notebook-app-component/README.md
@@ -1,0 +1,20 @@
+# Notebook Preview
+Preview Jupyter notebooks in a small React component.
+
+## Installation
+```
+npm install --save @nteract/notebook-preview
+```
+
+## Usage
+```jsx
+import NotebookPreview from "@nteract/notebook-preview";
+
+<NotebookPreview
+  notebook={myNotebook}
+/>
+```
+
+## See It in Action
+
+A demo of `notebook-preview` can be found in the [showcase](../showcase/pages/notebook-preview.js).
