+++ packages/notebook-app-component/__tests__/draggable-cell-spec.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import { shallow } from "enzyme";
 
-import DraggableCell from "../../src/components/draggable-cell";
+import DraggableCell from "../src/draggable-cell";
 
 import { emptyMarkdownCell } from "@nteract/commutable";
