+++ packages/notebook-app-component/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "@nteract/notebook-app-component",
+  "version": "1.0.0",
+  "description": "Editable notebook app component, backed by @nteract/core's store",
+  "main": "lib/index.js",
+  "nteractDesktop": "src/index.js",
+  "scripts": {
+    "prepare": "npm run build",
+    "prepublishOnly": "npm run build && npm run build:flow",
+    "build": "npm run build:clean && npm run build:lib",
+    "build:clean": "rimraf lib",
+    "build:flow": "flow-copy-source -v -i '**/__tests__/**' src lib",
+    "build:lib": "babel -d lib src --ignore '**/__tests__/**'",
+    "build:lib:watch": "npm run build:lib -- --watch",
+    "build:watch": "npm run build:clean && npm run build:lib:watch && npm run build:flow"
+  },
+  "repository": "https://github.com/nteract/nteract/tree/master/packages/notebook-app-component",
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "date-fns": "^1.29.0",
+    "@nteract/core": "^5.0.1",
+    "@nteract/commutable": "^4.0.1",
+    "@nteract/display-area": "^4.1.1",
+    "@nteract/markdown": "^1.1.0",
+    "@nteract/mathjax": "^1.0.0",
+    "@nteract/presentational-components": "^0.0.1",
+    "@nteract/transforms": "^4.1.1",
+    "@nteract/commutable": "^4.0.1",
+    "@nteract/display-area": "^4.1.1",
+    "@nteract/dropdown-menu": "^0.1.2",
+    "@nteract/editor": "^7.0.4",
+    "@nteract/markdown": "^1.1.0",
+    "@nteract/mathjax": "^1.0.0",
+    "@nteract/messaging": "^4.0.0",
+    "@nteract/octicons": "^0.2.0",
+    "@nteract/presentational-components": "^0.0.1",
+    "@nteract/transforms": "^4.1.1",
+    "react-dnd": "^2.5.4",
+    "react-dnd-html5-backend": "^2.5.4",
+    "react-redux": "^5.0.7",
+    "react-syntax-highlighter": "^7.0.1",
+    "redux": "^4.0.0"
+  },
+  "peerDependencies": {
+    "immutable": "^4.0.0-rc.9",
+    "react": "^16.2.0",
+    "styled-jsx": "^2.2.6"
+  },
+  "author": "Kyle Kelley <rgbkrk@gmail.com>",
+  "license": "BSD-3-Clause"
+}
