+++ packages/notebook-app-component/src/hijack-scroll.js
