+++ packages/notebook-app-component/src/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+import NotebookApp from "./notebook-app";
+export default NotebookApp;
